# openapi.yaml
openapi: 3.0.3
info:
  title: Epic Partners Тестовое задание
  description: Выполняю тестовое задание на вакансию "Junior backend разработчик (удаленно)"
  version: 0.0.1
servers:
  - url: https://localhost:8080/api/v1
    description: Dev Server
paths:
  /count/{id}:
    get:
      summary: Метод получения счетчика
      tags:
        - Counter
      operationId: getCount
      parameters:
        - name: id
          in: path
          required: true
          description: ID счетчика
          schema:
            type: string
      responses:
        '200':
          description: Вернул значение счетчика
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Counter"
        '404':
          description: Счетчик отсутствует
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Counter"
  /count:
    post:
      summary: Метод получения счетчика
      tags:
        - Counter
      operationId: incrementCount
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Counter"
      responses:
        '200':
          description: Успешно увеличил значение счетчика
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Counter"
        '400':
          description: Не задан counterId или incrementCount. Или если incrementCount не int или не больше 0
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Counter"
        '404':
          description: Счетчик отсутствует
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Counter"
            
            
components:
  schemas:
    Counter:
      type: object
      required: 
        - id
        - count
      properties:
        id:
          type: integer
          example: 1
        count:
          type: integer
          example: 420
          